Update gradle plugin version
This went crazy with 1.x upgrade to Android Studio, apparently
for no reason. I don't know why we should specify it more
carefully than this.

The actual error was
The project is using an unsupported version of the Android Gradle plug-in
which is really hard to track down to just a classpath in the
build.gradle!