Added adjustResize to HostContactActivity to address #244
With adjustResize, the activity's layout gets resized automatically
once the software keyboard is being shown. By doing this, the message
button is being pushed upwards.

With a 5" smartphone this works quite well as there is still enough
space for six lines of message text. However, on smaller devices this
might lead to undesirable behaviour.