Move hidden send button in HostContactActivity (#250)
The send button was hidden by the software keyboard. With adjustResize,
the activity's layout gets resized automatically once the keyboard is
being shown. By doing this, the send button is being pushed upwards.

With a 5" smartphone this works quite well as there is still enough
space for six lines of message text. However, on smaller devices this
might lead to undesirable behaviour.

Fixes:  #244
Closes: #250