Eventually close the host search progress dialog (#251)
Changing the orientation of the device while a host search was running
resulted in a progress dialog that was never closed. The app needed to
be killed.

This is fixed by switching to a DialogFragment for the progress dialog
as well as saving the dialog state prior to the orientation change and
restoring from it after.

Currently, only ListSearchTabActivity has been moved to using that
DialogFragment. All other activities that might show a progress dialog
still use the old DialogHandler version.

The HostListAdapter now also has a resetDataset method which accepts a
new list, clears the old dataset, sets the specified list as the new
one, and notifies itself of the change in order to update the UI.

Also, the string "search_results" that spanned 2 activities is now
defined in a constant.

Fixes:  #243
Closes: #251