Fixes issue #243 by saving the dialogue state
Currently, this fix only works in cases where the search has already
been successful before the device is being rotated.

Besides fixing the issue, I made the host list adapter a member var.
The adapter now also has a resetDataset method which accepts a new List,
clears the old dataset, sets the specified list as a new one, and
notifies itself of a new dataset in order to change the UI.