Make the tests green again
All the tests failed. The test resources directory was wrong (or
outdated) and needed to be moved from src/test/res/ to
src/test/resources/. The mock server replies were not reflecting real
ones and missed some properties. And a few test cases were not updated
when code was changed in the past.