Make TLS SNI work
warmshowers.org switched to a hoster which uses SNI. The Apache HTTPClient does
not support this out of the box and the app is therefore no longer able to log
in. However, a fix on devices running Android 4.2+ is possible by using the
native hostname verifier. For Android 4.1 and earlier it is possibly supported
through reflexion on the same methods.