Make the project building and green again (#255)
* Use legacy http library

The org.apache.http library has been deprecated. It is now available as
org.apache.http.legacy.

* Switch to newer library versions

Gradle, the build tools, SDK and many libraries are brought to newer
versions (as far as compatible).

* Fix all the lint errors and some of the warnings

* Make the tests compile again

Adds jUnit as a testCompile dependency (thank you @PattaFeuFeu) and
fixes outdated calls to the MapSearchJsonParser constructor.

* Make the tests green again

All the tests failed. The test resources directory was wrong (or
outdated) and needed to be moved from src/test/res/ to
src/test/resources/. The mock server replies were not reflecting real
ones and missed some properties. And a few test cases were not updated
when code was changed in the past.

* Set a default debug keystore

Adds a keystore for signing debug apk files. This allows installation
of arbitrary debug versions without the hassle of dealing with
different signing keys. The release keystore settings have to be set in
local.properties.