Make TLS SNI work (#256)
warmshowers.org switched to a hoster which uses SNI. On Android < 6.0
the Apache HTTPClient does not support this out of the box. Therefore,
the app is no longer able to verify connections with the server and
thus fails to operate. However, a fix on devices running Android 4.2+
is possible by using the native hostname verifier. For Android 4.1 and
earlier it is possibly supported through reflexion on the same methods.