Fix performance penalty in String processing (#259)
The method defines a String object and loops through the hosts to add
their info to the String, under the hood the String is converted to a
StringBuffer/StringBuilder, appended to, and converted back to a String.
This can lead to a cost quadratic in the number of iterations, as the
growing string is recopied in each iteration. This commit replaces
String with StringBuilder and convert it back to the String object when
the string process is over.