Switch authentication mechanism to retrofit2
Switches the authentication logic to use the retrofit2 library. It also
introduces RxAndroid2 for a reactive way of passing updates in the app.

The changes are transparently introduced s.t. the old RestClient can
still operate. It is going to be removed in future commits.

Logging out from an account is currently broken but is eventually going
to be dropped anyways (logging out is really just removing the account).