Replace activities by fragments
Activities required a lot of duplicated code to handle the same things
as e.g. the navigation drawer. Having a main activity which loads
fragments for the different views is a much cleaner approach.

This change touches almost all frontend code and a lot of cleanup and
refactoring is done in the same time than switching to fragments.