Refactor the backend
The database is split into two parts - one which contains data for all
WS accounts and one which is account specific (this will eventually
allow us to have multiple accounts).

Implements a cache layer (repository) which connects the database and
the network. It reduces network accesses and also introduces the
possibility to decide if our data from the db is fresh enough and does
not need to be updated from the network (TBD).

Fetching of users and feedbacks is now switched to retrofit2.

Removes the deprecated HostBriefInfo and StarredHostDao classes.

Some other minor refactorings and bugfixes.