Add initial version of message notifications
Adds a service that periodically checks for new messages and displays
a notification to the user as soon as one is received.

Also, introduces the concept of a pushable. This is a type that needs to
be sent to the webservice, however needs to be stored locally as long as
said update did not go successfully through yet (e.g. the message's read
status or a new message itself).