Improve location handling
Improves the way the initial location on the map is fetched and
handled. Uses the logged in user's home address as the initial location
if none is saved from a previous run. In case no such address is
available a per-language default location is used. This is now switched
to a more zoomed out view as well as a Europe-centered default for
languages that are different from english.

Also replaces the deprecated FusedLocationApi by the
FusedLocationProviderClient. This makes code easier as we are no longer required
to manually do the connection handling with the Google Play service.