Refactor code that depends on a logged in user
Makes use of dagger subcomponents to have seperate state for each
logged in user. This is useful for when we want to allow to have more
than one account in the app. But already now it resolves a lot of bugs
in case an account is removed at any time. And it improves the
readability of the code as dependencies are much more explicit.
Moreover, it makes the story of "what is available when?" much easier to
understand.