Add reactive way to handle activity results
If no user is logged in or its auth tokens are outdated we sometimes
have to wait for the main activity to be around until can switch to the
AuthActivity. This is now properly implemented to work in a reactive
way. This solves an issue when migrating from the 1.5 version where the
app was only usable after a restart because we were not able to react to
the new auth tokens to become available.