Updates compileSdk, sets dynamic versionName/versionCode
In order to not always having to change the versionName, this includes a
calculation. The versionName is concatenated based off of versionMajor,
versionMinor, and versionPatch variables.

versionCode is calculated by multiplying those three in addition to
versionBuild, thus leading to a unique versionCode each time one of the
variable gets an increase.

For betas, such as "-beta2", just increase "versionBuild" to 1,2,..