Decouple Rx callbacks from their UI handlers
Decouple callbacks that are processed upon arrival from their handlers
that can only be executed when the app is in the foreground.

Also fixes a bug in MessageThreadsFragment where Observable::buffer()
was misused to debounce burst updates to the message thread list to
avoid flickering. However, this method issues the callback at a fixed
time interval, also when no new data is around, and so a lot of updates
were made all the time (with no effect).