Fix race in database open
The account database handler used to react to account changes and
(re-)open the database for the new account. This imposed two errors:
First, the handler is in the account scope and should open the
connection to only its account db. Second, the reaction code was async
and the db ended up to sometimes not being opened when the first access
already happened.

This also replaces the initial lock in LocableDatabase by an explicit
check whether there is an open db connection upon execution of a query.