Make expired auth token renewals reentrant
If multiple API requests were in-flight and failing because of an
expired auth token, that auth token was marked as expired multiple times
which resulted in the login activity being shown for each such call.

Also, the AuthenticationController is now fully per-account. It was
already in the account scope but always handled the current account.