Add user cache -- (#292)
Users that were loaded from the network are now kept in a cache and only
discarded when the app is closed. This means, that if a user navigates
away from the map and then comes back to it no more loading needs to be
done for the previously loaded areas.
Also, the loading is area aware. That means that for overlapping areas
requests are now only made for the non-loaded parts.

Fixes: #101