Fix issue where DB migration ran twice
In case the account db does not yet exist it is created in the most
up-to-date version. Then the migration to version 1 needs to be done to
recover the favorite users when updating from a version where the app
and account databases were not yet split. However, the rest of the
migrations must not be done as they are already applied when creating
the db.