Immediately put the logged in user into cache
The user resource is returned in the successful login reply from the
server. We now immediately put it into cache s.t. it is always around
and does not need any further network accesses to fetch it.

This fixes an issue where on a fresh install the map was not centered
on the logged in users' home address in case a current location is not
known from any location provider. This was due to the user resource of
the logged in account not being around when the map was first centered.