Cleanup code that works around osmdroid's multi callback issues
In osmdroid there is no way of setting the center and the zoom of the
map at the same time. When using two separate calls to `setCenter()` and
`setZoom()` then the callbacks of the current center/zoom are called
multiple times, some with old attributes on the map.
This is a cleanup of the code that works around those issues.