Fix bug in Repository
If an error occured while fetching an entry that is kept in a
repository, its cache entry status was not updated to 'LOADING' after
issuing a second load attempt. Subscribers that newly attached to it
then initially saw the erroneous status from the last loading attempt.

That resulted in the UserFragment being closed even though the load
succeeded after it failed the last time.