Add the option to remember the user's password (#312)
The auth token that is used to authenticate the user with the
Warmshowers webservice times out after about one week of not using it.
After that, the user needs to re-login. We decided a while back to no
longer store the user's password on the device, however this constant
need to re-login is an inconvenience. Therefore, there is now a
checkbox to let the user decide whether they want to save the password.

This change also contains some modifications to make the app more
testable and mockable.

Fixes: #310
Closes: #312