Improve tests for Message
This divides the Robolectric tests from the non-Robolectric tests in
MessageTest to improve the speed and have them more side-effect free.
On my (old) machine, this sped up the tests from 16s runtime to 10-11s.

Also, it fixes an issue where I didn’t compare the right parameters when
testing for the clone methods.

Lastly, this adds other test cases for the parseBody method.