Always reload user when the UserFragment resumes
We used to load the user resource only at fragment creation. However,
if one provided new feedback for the user through the app and then
returned to the user fragment the new feedback was not shown.
Since the user resource is cached anyways we can just always reload it
and immediately show potentially new data.