Refactor message notifications
Differentiate between "noticed" and "read" messages. A user might
notice that a new message arrived but decide to not yet read it. Yet,
we are no longer showing any notifications for noticed messages.

The notification controller is also refactored to use
`NotificationCompat` on all devices. Also, a notification channel is
created on Android O+ devices.